requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
anyio==3.7.1
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
Babel==2.12.1
bcrypt==4.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
cssselect==1.2.0
fastapi==0.99.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
lxml==4.9.3
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
pydantic==1.10.11
pyquery==2.0.0
python-dotenv==1.0.0
pytz==2023.3
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
yarl==1.9.2
